pyproject.toml
[project]
name = "fa-dae2-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "psycopg[binary]~=3.2",
    "polars>=1.2",
    "snowflake-connector-python~=3.16",
    "faker>=25",
    "confluent-kafka~=2.4",
    "openai>=1.99.9",
    "pypdf2>=3.0.1",
    "pinecone>=7.3.0",
]

[dependency-groups]
ci = [
    "dbt-snowflake>=1.10.0",
    "sqlfluff>=3.4.2",
    "sqlfluff-templater-dbt>=3.4.2",
]
dev = [
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.4",
]

[tool.ruff]
ignore = ["E402"]
